% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bound_frequency_exceeding_bp.R
\name{bound_frequency_exceeding_bp}
\alias{bound_frequency_exceeding_bp}
\title{This function uses the Nelder-Mead algorithm  for the optimization procedure. This algorithm
is implemented in the 'nmkb' function from the {dfoptim} package.}
\usage{
bound_frequency_exceeding_bp(
  obj_func_bp,
  maximize = FALSE,
  lower_parameters = c(1, -5, -20),
  upper_parameters = c(6, 1, -10),
  niter_ale = 1000,
  niter_epi = 1000,
  threshold = 1,
  percentile_ale = NULL,
  suff_stat_concentration,
  suff_stat_consumption,
  consumption_change_vals_EKE = c(-15, 7.5),
  consumption_change_probs_EKE = c(0.25, 0.75),
  consumers_info_sample_size,
  concentration_mu0 = 2.75,
  concentration_v0 = 5,
  concentration_alpha0 = 1,
  concentration_beta0 = 1,
  sufficient_statistics_concentration = TRUE,
  consumption_mu0 = -2.5,
  consumption_v0 = 5,
  consumption_alpha0 = 1,
  consumption_beta0 = 1,
  sufficient_statistics_consumption = TRUE,
  consumption_event_alpha0 = 1,
  consumption_event_beta0 = 1
)
}
\arguments{
\item{obj_func_bp}{The objective function to optimize.}

\item{maximize}{A logical variable indicating whether the objective function should be maximized. Default is \code{FALSE}.}

\item{lower_parameters}{Lower bounds on the parameters. A vector of the same length as the parameters.}

\item{upper_parameters}{Upper bounds on the parameters. A vector of the same length as the parameters.}

\item{niter_ale}{number of generated samples}

\item{niter_epi}{number of generated parameters from the posterior distrbutions
(it indicates the number of repetitions the assessment will be done)}

\item{threshold}{safety threshold}

\item{percentile_ale}{a value that indicates if the assessment is done on a random child  by \code{NULL} or on a high consumer child by 95. Default is NULL}

\item{suff_stat_concentration}{a vector of sufficient statistics: sample_size, sample_mean and sample_sd
corresponding to concentration. If sufficient_statistics_concentration = \code{FALSE},
then it is vector of observed data}

\item{suff_stat_consumption}{a vector of sufficient statistics: sample_size, sample_mean and sample_sd
corresponding to consumption. If sufficient_statistics_consumption = \code{FALSE},
then it is vector of observed data}

\item{consumption_change_vals_EKE}{a vector of elicited values from experts}

\item{consumption_change_probs_EKE}{a vector of elicited probabilities from experts}

\item{consumers_info_sample_size}{a vector with the sample size of non_consumer_sample_size and consumer_sample_size}

\item{concentration_mu0}{prior hyperparameter \emph{mu0} for the normal-gamma distribution corresponding to concentration}

\item{concentration_v0}{prior hyperparameter \emph{v0} for the normal-gamma distribution corresponding to concentration}

\item{concentration_alpha0}{prior hyperparameter \emph{alpha0} for the normal-gamma distribution  corresponding to concentration}

\item{concentration_beta0}{prior hyperparameter \emph{beta0} for the normal-gamma distribution corresponding to concentration}

\item{sufficient_statistics_concentration}{logical; if \code{TRUE}, sufficient statistics (sample_size, sample_mean, sample_variance)
corresponding to concentration are given as input data, otherwise
sufficient_statistics_concentration is given as observed data. Default is \code{TRUE}}

\item{consumption_mu0}{prior hyperparameter \emph{mu0} for the normal-gamma distribution corresponding to consumption}

\item{consumption_v0}{prior hyperparameter \emph{v0} for the normal-gamma distribution corresponding to consumption}

\item{consumption_alpha0}{prior hyperparameter \emph{alpha0} for the normal-gamma distribution corresponding to consumption}

\item{consumption_beta0}{prior hyperparameter \emph{beta0} for the normal-gamma distribution corresponding to consumption}

\item{sufficient_statistics_consumption}{logical; if \code{TRUE}, sufficient statistics (sample_size, sample_mean, sample_variance)
corresponding to consumption are given as input data, otherwise
sufficient_statistics_consumption is given as observed data. Default is \code{TRUE}}

\item{consumption_event_alpha0}{prior hyperparameter \emph{alpha0} for the beta distribution corresponding to consumption event}

\item{consumption_event_beta0}{prior hyperparameter \emph{beta0} for the beta distribution corresponding to consumption event}
}
\value{
## Two lists
\enumerate{
\item opt_value
\item opt_frequency
}
The components of the first list (\emph{opt_value}) are
\describe{
\item{par}{Best estimate of the parameter vector found by the algorithm}
\item{value}{The value of the objective function at termination}
\item{feval}{The number of times the objective function was evaluated}
\item{restarts}{The number of times the algorithm had to be restarted when it stagnated}
\item{convergence}{An integer code indicating type of convergence. 0 indicates successful convergence. Positive integer codes indicate failure to converge}
\item{message}{Text message indicating the type of convergence or failure}
}
The components of the second list (\emph{opt_prob}) are
\describe{
\item{prob_consumption_event}{The estimated probability of consumption events}
\item{parameters_concentration}{A list with the values of the prior and posterior parameters of concentration}
\item{parameters_consumption}{A list with the prior and posterior parameters of consumption}
\item{frequency_exceeding}{A vector with the estimated frequency of exceeding the threshold (the lenght is niter_epi)}
\item{expected_frequency_exceeding}{The expected value of the frequency of exceeding the threshold}
\item{hdi_frequency_exceeding}{The highest posterior density interval of the frequency of exceeding the threshold}
}
}
\description{
This function uses the Nelder-Mead algorithm  for the optimization procedure. This algorithm
is implemented in the 'nmkb' function from the {dfoptim} package.
}
