% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_uncertainty.R
\name{combine_uncertainty}
\alias{combine_uncertainty}
\title{Combine uncertainty}
\usage{
combine_uncertainty(
  gen_data_concentration,
  gen_data_consumption,
  gen_data_EKE,
  threshold,
  niter_ale
)
}
\arguments{
\item{gen_data_concentration}{generated concentration samples after Bayesian inference (output of function generate_samples_normal_gamma)}

\item{gen_data_consumption}{generated consumption samples after Bayesian inference (output of function generate_samples_normal_gamma)}

\item{gen_data_EKE}{a generated sample from the fitted distribution to the expert's elicited values}

\item{threshold}{safety threshold}

\item{niter_ale}{number of generated samples}
}
\value{
prob_exceed_wi           the probability of exceeding the weekly intake
}
\description{
This function estimates the probability of exceeding the weekly intake
}
