% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/estimate_parameters_normal_dist_from_two_percentiles.R
\name{estimate_parameters_normal_dist_from_two_percentiles}
\alias{estimate_parameters_normal_dist_from_two_percentiles}
\title{Fit a normal distribution using two quantiles}
\usage{
estimate_parameters_normal_dist_from_two_percentiles(vals, probs)
}
\arguments{
\item{vals}{A vector of elicited values}

\item{probs}{A vector of elicited probabilities}
}
\value{
A vector with the parameters mean (\emph{mu_EKE}) and standard deviation (\emph{sigma_EKE}) of the fitted normal distribution.
}
\description{
Estimate the parameters of a normal distribution using two quantiles
}
\examples{
vals = c(-30, -15, 05, 7.5, 20)
probs = c(0.01, 0.25,0.5, 0.75,0.99)
estimate_parameters_normal_dist_from_two_percentiles(vals, probs)

}
