% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_bp.R
\name{graph_bp}
\alias{graph_bp}
\title{Graph - Bounded probability}
\usage{
graph_bp(lower_points, upper_points)
}
\arguments{
\item{lower_points}{A vector with the lower cdf sequence}

\item{upper_points}{A vector with the upper cdf sequence}
}
\value{
a plot
}
\description{
Plot a p-box (a p-box is represented by two cdf sequences)
}
