% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_bp_both.R
\name{graph_bp_both}
\alias{graph_bp_both}
\title{Graph both - Bounded probability}
\usage{
graph_bp_both(
  lower_points_average_consumption,
  upper_points_average_consumption,
  lower_points_high_consumption,
  upper_points_high_consumption
)
}
\arguments{
\item{lower_points_average_consumption}{A vector with the lower cdf sequence corresponding to average consumption scenario}

\item{upper_points_average_consumption}{A vector with the upper cdf sequence corresponding to average consumption scenario}

\item{lower_points_high_consumption}{A vector with the lower cdf sequence corresponding to high consumption scenario}

\item{upper_points_high_consumption}{A vector with the upper cdf sequence corresponding to high consumption scenario}
}
\value{
a plot
}
\description{
Plot two p-boxes (each p-box is represented by two cdf sequences)
All inputs much have the same length
}
