% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_pp.R
\name{graph_pp}
\alias{graph_pp}
\title{Graph - Precise probability}
\usage{
graph_pp(prob_exceed)
}
\arguments{
\item{prob_exceed}{a vector of probabilities of exceedance}
}
\value{
a plot
}
\description{
Plot a cumulative distribution function.
}
