% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective_function_bp.R
\name{obj_func_bp}
\alias{obj_func_bp}
\title{Objective function for Bounded probability}
\usage{
obj_func_bp(
  parameters,
  niter_ale,
  niter_epi,
  threshold,
  exposure_scenario,
  suff_stat_concentration,
  suff_stat_consumption,
  consumption_change_vals_EKE,
  consumption_change_probs_EKE,
  consumers_info_sample_size,
  concentration_v0,
  concentration_alpha0,
  concentration_beta0,
  sufficient_statistics_concentration,
  consumption_v0,
  consumption_alpha0,
  consumption_beta0,
  sufficient_statistics_consumption,
  consumption_event_alpha0,
  consumption_event_beta0,
  percentile = NULL
)
}
\arguments{
\item{parameters}{parameters of the objective function (concentration_mu0, consumption_mu0, consumption_change_vals_EKE)}

\item{niter_ale}{number of generated samples}

\item{niter_epi}{number of generated parameters from the posterior distrbutions
(it indicates the number of repetitions the assessment will be done)}

\item{threshold}{safety threshold}

\item{exposure_scenario}{a value that indicates if the assessment is done on an average consumption scenario by 'av' or on a high consumption scenario by 'perc_95'. Default is 'av'}

\item{suff_stat_concentration}{a vector of sufficient statistics: sample_size, sample_mean and sample_sd
corresponding to concentration. If sufficient_statistics_concentration = \code{FALSE},
then it is vector of observed data}

\item{suff_stat_consumption}{a vector of sufficient statistics: sample_size, sample_mean and sample_sd
corresponding to consumption. If sufficient_statistics_consumption = \code{FALSE},
then it is vector of observed data}

\item{consumption_change_vals_EKE}{a vector of elicited values from experts}

\item{consumption_change_probs_EKE}{a vector of elicited probabilities from experts}

\item{consumers_info_sample_size}{a vector with the sample size of non_consumer_sample_size and consumer_sample_size}

\item{concentration_v0}{prior hyperparameter \emph{v0} for the normal-gamma distribution corresponding to concentration}

\item{concentration_alpha0}{prior hyperparameter \emph{alpha0} for the normal-gamma distribution  corresponding to concentration}

\item{concentration_beta0}{prior hyperparameter \emph{beta0} for the normal-gamma distribution corresponding to concentration}

\item{sufficient_statistics_concentration}{logical; if \code{TRUE}, sufficient statistics (sample_size, sample_mean, sample_variance)
corresponding to concentration are given as input data, otherwise
sufficient_statistics_concentration is given as observed data. Default is \code{TRUE}}

\item{consumption_v0}{prior hyperparameter \emph{v0} for the normal-gamma distribution corresponding to consumption}

\item{consumption_alpha0}{prior hyperparameter \emph{alpha0} for the normal-gamma distribution corresponding to consumption}

\item{consumption_beta0}{prior hyperparameter \emph{beta0} for the normal-gamma distribution corresponding to consumption}

\item{sufficient_statistics_consumption}{logical; if \code{TRUE}, sufficient statistics (sample_size, sample_mean, sample_variance)
corresponding to consumption are given as input data, otherwise
sufficient_statistics_consumption is given as observed data. Default is \code{TRUE}}

\item{consumption_event_alpha0}{prior hyperparameter \emph{alpha0} for the beta distribution corresponding to consumption event}

\item{consumption_event_beta0}{prior hyperparameter \emph{beta0} for the beta distribution corresponding to consumption event}

\item{percentile}{a value between 1 and 100 which indicates a percentile. By default
is NULL}
}
\value{
expected_frequency_exceeding               the expected value of the frequency of exceeding the threshold
}
\description{
This is the objective function (a function to be optimized)
The objective function is the unc_analysis_assessment_bp function where
concentration_mu0, consumption_mu0 and consumption_change_vals_EKE
are the parameters and the rest of the inputs arguments are fixed.
}
