% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify_uncertainty_pp_change_eke.R
\name{quantify_uncertainty_pp_change_eke}
\alias{quantify_uncertainty_pp_change_eke}
\title{Fit a Normal distribution from Expert Kwonledge elicitation}
\usage{
quantify_uncertainty_pp_change_eke(vals, probs)
}
\arguments{
\item{vals}{A vector of elicited values for one expert, or a matrix of elicited values for multiple
experts (one column per expert). Note that the an elicited judgement about X should be of the form P(X<= vals[i,j]) = probs[i,j]}

\item{probs}{A vector of elicited probabilies for one expert, or a matrix of elicited values for multiple
experts (one column per expert). A single vector can be used if the probabilities are the same
for each expert. For each expert, the smallest elicited probability must be less than 0.4,
and the largest elicited probability must be greater than 0.6.}
}
\value{
A vector with the parameters mean (\emph{mu_EKE}) and standard deviation (\emph{sigma_EKE}) of the fitted normal distribution.
}
\description{
Takes elicited probabilities as inputs, and fits a normal distribution using least squares on
the cumulative distribution function. If separate judgements from multiple experts are specified,
the function will fit one set of distributions per expert.
}
\examples{
vals = c(-30, -15, 05, 7.5, 20)
probs = c(0.01, 0.25,0.5, 0.75,0.99)
quantify_uncertainty_pp_change_eke(vals, probs)

}
