% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_bernoulli_beta.R
\name{update_bernoulli_beta}
\alias{update_bernoulli_beta}
\title{Bernoulli-Beta conjugate model}
\usage{
update_bernoulli_beta(suff_stat_data, alpha0 = 1, beta0 = 1)
}
\arguments{
\item{suff_stat_data}{a vector of sufficient statistics: non_consumer_sample_size and consumer_sample_size}

\item{alpha0}{prior hyperparameter \emph{alpha0} for the beta distribution}

\item{beta0}{prior hyperparameter \emph{beta0} for the beta distribution}
}
\value{
param  A list with the prior and posterior hyperparameters. It constains the following components
\describe{
\item{prior}{A list with the prior hyperparameters \emph{alpha, beta}}
\item{posterior}{A list with the posterior hyperparameters \emph{alpha, beta}}
}
}
\description{
This function implements the conjugate model (B | pi) ~ Bernoulli(pi); pi ~ Beta(alpha0,beta0)
See wikipedia for details https://en.wikipedia.org/wiki/Conjugate_prior
}
\examples{
data <- c(non_consumer_sample_size = 20, consumer_sample_size = 10)
update_bernoulli_beta(suff_stat_data = data, alpha0 = 1, beta0 = 1)

}
