% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_normal_gamma.R
\name{update_normal_gamma}
\alias{update_normal_gamma}
\title{Conjugate Model Normal - Normal Gamma}
\usage{
update_normal_gamma(
  suff_stat_data,
  mu0,
  v0,
  alpha0,
  beta0,
  sufficient_statistics
)
}
\arguments{
\item{suff_stat_data}{a vector of sufficient statistics: sample_size, sample_mean and sample_sd. If sufficient_statistics = \code{FALSE}, then it is vector of observed data}

\item{mu0}{prior hyperparameter \emph{mu0} for the normal-gamma distribution}

\item{v0}{prior hyperparameter \emph{v0} for the normal-gamma distribution}

\item{alpha0}{prior hyperparameter \emph{alpha0} for the normal-gamma distribution}

\item{beta0}{prior hyperparameter \emph{beta0} for the normal-gamma distribution}

\item{sufficient_statistics}{logical; if \code{TRUE}, sufficient statistics, suff_stat_data, are given as input data, otherwise suff_stat_data is given as observed data. Default is \code{TRUE}}
}
\value{
param  A list with the prior and posterior hyperparameters. It constains the following components

\describe{
  \item{prior}{A list with the prior hyperparameters \emph{mu, v, alpha, beta}}

  \item{posterior}{A list with the posterior hyperparameters \emph{mu, v, alpha, beta}}
}
}
\description{
The conjugate model is
(x| mu,tau) ~ Normal(mu, tau);
(mu, tau) ~ Normal-Gamma(mu0,v0,alpha0,beta0)
See wikipedia for details https://en.wikipedia.org/wiki/Conjugate_prior
}
\examples{
dta <- rnorm(100)
update_normal_gamma(suff_stat_data = dta, mu0 = 0,
                   v0 = 5, alpha0 = 1, beta0 = 1, sufficient_statistics = FALSE)

}
