% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_bp_both.R
\name{graph_bp_both}
\alias{graph_bp_both}
\title{Graph both - Bounded probability}
\usage{
graph_bp_both(
  lower_points_all,
  upper_points_all,
  lower_points_high_consumer,
  upper_points_high_consumer
)
}
\arguments{
\item{lower_points_all}{A vector with the lower cdf sequence corresponding to all children}

\item{upper_points_all}{A vector with the upper cdf sequence corresponding to all children}

\item{lower_points_high_consumer}{A vector with the lower cdf sequence corresponding to a high consumer child}

\item{upper_points_high_consumer}{A vector with the upper cdf sequence corresponding to a high consumer child}
}
\value{
a plot
}
\description{
Plot two p-boxes (each p-box is represented by two cdf sequences)
All inputs much have the same length
}
