% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_pp.R
\name{graph_pp}
\alias{graph_pp}
\title{Graph - Precise probability}
\usage{
graph_pp(frequency_exceeding)
}
\arguments{
\item{frequency_exceeding}{a vector of frequency of exceeding}
}
\value{
a plot
}
\description{
Plot a cumulative distribution function.
}
