% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_pp_both.R
\name{graph_pp_both}
\alias{graph_pp_both}
\title{Graph both - Precise probability}
\usage{
graph_pp_both(
  frequency_exceeding_random_child,
  frequency_exceeding_high_consumer
)
}
\arguments{
\item{frequency_exceeding_random_child}{a vector of frequency of exceedance corresponding to a random child}

\item{frequency_exceeding_high_consumer}{a vector of probabilities of exceedance corresponding to a high consumer child}
}
\value{
a plot
}
\description{
Plot two cumulative distribution functions: a random child  and a high consumer child.
frequency_exceeding_random_child and frequency_exceeding_high_consumer must have the same length
}
