% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_pp_both.R
\name{graph_pp_both}
\alias{graph_pp_both}
\title{Graph both - Precise probability}
\usage{
graph_pp_both(prob_exceed_all, prob_exceed_high_consumer)
}
\arguments{
\item{prob_exceed_all}{a vector of probabilities of exceedance corresponding to all children}

\item{prob_exceed_high_consumer}{a vector of probabilities of exceedance corresponding to a high consumer child}
}
\value{
a plot
}
\description{
Plot two cumulative distribution functions: all children and a high consumer child.
prob_exceed_all and prob_exceed_high_consumer must have the same length
}
