% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_pp_both.R
\name{graph_pp_both}
\alias{graph_pp_both}
\title{Graph both - Precise probability}
\usage{
graph_pp_both(
  frequency_exceeding_average_consumption,
  frequency_exceeding_high_consumption
)
}
\arguments{
\item{frequency_exceeding_average_consumption}{A vector of frequency of exceeding corresponding to average consumption}

\item{frequency_exceeding_high_consumption}{A vector of frequency of exceeding corresponding to high consumption}
}
\value{
a plot
}
\description{
Plot two cumulative distribution functions: average consumption and high consumption
frequency_exceeding_average_consumption and frequency_exceeding_high_consumption must have the same length
}
